Updated admin functionality quite a bit

+++ Pages/Admin.cshtml.cs
@@ -44,7 +44,7 @@ namespace TannersWebsiteTemplate.Pages
 
         public async Task<IActionResult> OnPostDelete(int? userid)
         {
-            if (HttpContext.Session.GetInt32("IsAdmin") == 1 && SQL.Admin.IsAdmin(HttpContext.Session.GetInt32("UserId"))) // is the user admin and does the userid in session check out as admin? if so, delete user requested.
+            if (HttpContext.Session.GetInt32("IsAdmin") == 1 && SQL.Admin.IsAdmin(HttpContext.Session.GetInt32("UserId")) && HttpContext.Session.GetInt32("UserId") == 1) // is the user admin and is it super admin doing this and does the userid in session check out as admin? if so, delete user requested.
             {
                 await SQL.Admin.DeleteUser(userid);
             }
@@ -53,7 +53,7 @@ namespace TannersWebsiteTemplate.Pages
         }
         public async Task<IActionResult> OnPostAdmin(int? userid)
         {
-            if (HttpContext.Session.GetInt32("IsAdmin") == 1 && HttpContext.Session.GetInt32("UserId") == 1 && SQL.Admin.IsAdmin(HttpContext.Session.GetInt32("UserId"))) // is the user admin and is it super admin doing this and does the userid in session check out as admin? if so, make the requested userid an admin
+            if (HttpContext.Session.GetInt32("IsAdmin") == 1 && SQL.Admin.IsAdmin(HttpContext.Session.GetInt32("UserId")) && HttpContext.Session.GetInt32("UserId") == 1) // is the user admin and is it super admin doing this and does the userid in session check out as admin? if so, make the requested userid an admin
             {
                 await SQL.Admin.AdminUser(userid);
             }
@@ -72,7 +72,7 @@ namespace TannersWebsiteTemplate.Pages
 
         public async Task<IActionResult> OnPostUnbanUser(int? userid)
         {
-            if (HttpContext.Session.GetInt32("IsAdmin") == 1 && HttpContext.Session.GetInt32("UserId") == 1 && SQL.Admin.IsAdmin(HttpContext.Session.GetInt32("UserId"))) // is the user admin and is it super admin doing this and does the userid in session check out as admin? if so, make the requested userid an admin
+            if (HttpContext.Session.GetInt32("IsAdmin") == 1 && SQL.Admin.IsAdmin(HttpContext.Session.GetInt32("UserId"))) // is the user admin and does the userid in session check out as admin? if so, make the requested userid an admin
             {
                 await SQL.Admin.UnbanUser(userid);
             }
@@ -81,7 +81,7 @@ namespace TannersWebsiteTemplate.Pages
 
         public async Task<IActionResult> OnPostBanUser(int? userid)
         {
-            if (HttpContext.Session.GetInt32("IsAdmin") == 1 && HttpContext.Session.GetInt32("UserId") == 1 && SQL.Admin.IsAdmin(HttpContext.Session.GetInt32("UserId"))) // is the user admin and is it super admin doing this and does the userid in session check out as admin? if so, make the requested userid an admin
+            if (HttpContext.Session.GetInt32("IsAdmin") == 1 && SQL.Admin.IsAdmin(HttpContext.Session.GetInt32("UserId"))) // is the user admin and does the userid in session check out as admin? if so, make the requested userid an admin
             {
                 await SQL.Admin.BanUser(userid, "You have been banned.", DateTime.Now.AddYears(999));
             }
